termisignal/requirements.txt
textual>=0.38.1
pynacl>=1.5.0
websockets>=11.0.3
python-dotenv>=1.0.0
cryptography>=41.0.0
